+++ identity/backend/src/main/java/io/camunda/identity/Application.java
@@ -15,7 +15,7 @@ import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
  * This application file is a temporary addition to support the development of the minimal feature
  * set of Identity. It will be removed once the monojar integration is clear and available.
  */
-@SpringBootApplication(scanBasePackages = {"io.camunda.identity", "io.camunda.authentication"})
+@SpringBootApplication(scanBasePackages = {"io.camunda.authentication", "io.camunda.identity"})
 @ConfigurationPropertiesScan()
 public class Application {
