+++ identity/backend/src/main/java/io/camunda/identity/usermanagement/service/GroupService.java
@@ -8,7 +8,7 @@
 package io.camunda.identity.usermanagement.service;
 
 import io.camunda.authentication.user.CamundaUserDetailsManager;
-import io.camunda.identity.usermanagement.Group;
+import io.camunda.identity.user.Group;
 import io.camunda.identity.usermanagement.repository.GroupRepository;
 import java.util.Collections;
 import java.util.List;
