--- identity/backend/src/main/resources/scripts/users.ddl
@@ -1,11 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-create table users(id bigint generated by default as identity(start with 0) primary key, username varchar_ignorecase(50) not null unique,password varchar_ignorecase(500) not null,enabled boolean not null);
-create table profiles(id bigint primary key, email varchar_ignorecase(100) unique,constraint fk_profile_users foreign key(id) references users(id) ON DELETE CASCADE);
-create table authorities (username varchar_ignorecase(50) not null,authority varchar_ignorecase(50) not null,constraint fk_authorities_users foreign key(username) references users(username));
-create unique index ix_auth_username on authorities (username,authority);
