+++ identity/backend/src/test/java/io/camunda/identity/usermanagement/service/GroupServiceTest.java
@@ -11,7 +11,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import io.camunda.identity.usermanagement.Group;
+import io.camunda.identity.user.Group;
 import java.util.UUID;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
