+++ identity/common/src/main/java/io/camunda/identity/user/Group.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.identity.usermanagement;
+package io.camunda.identity.user;
 
 public record Group(Integer id, String name) {
   public Group(final String name) {
