+++ identity/common/src/main/resources/schema.sql
@@ -8,7 +8,8 @@
 
 -- Ref: https://docs.spring.io/spring-security/reference/servlet/appendix/database-schema.html#_user_schema
 CREATE TABLE IF NOT EXISTS users (
-    username varchar not null primary key,
+    id bigint generated by default as identity(start with 0) primary key,
+    username varchar_ignorecase(50) not null unique,
     password varchar not null,
     enabled boolean not null
 );
@@ -45,4 +46,6 @@ CREATE TABLE IF NOT EXISTS group_members (
     constraint fk_group_members_group foreign key(group_id) references groups(id),
     constraint uk_group_members_group_username unique(group_id , username)
 );
+
+create table profiles(id bigint primary key, email varchar_ignorecase(100) unique,constraint fk_profile_users foreign key(id) references users(id) ON DELETE CASCADE);
